Migrate SettingsCoordinator items (Indicators) to the CustomizeSettingsCoordinator. This is done with the goal of allowing Indicators + any future nested pages to be routable on tvOS.
